+++ src/capsule/react-native/ReactNativeCapsuleSigner.ts
@@ -1,0 +1,9 @@
+import { PrivateKeyStorage } from '../PrivateKeyStorage'
+import { CapsuleBaseSigner } from '../CapsuleSigner'
+import { ReactNativePrivateKeyStorage } from './ReactNativePrivateKeyStorage'
+
+export class ReactNativeCapsuleSigner extends CapsuleBaseSigner {
+  protected getPrivateKeyStorage(account: string): PrivateKeyStorage {
+    return new ReactNativePrivateKeyStorage(account)
+  }
+}
