+++ src/capsule/test/ChallengeStorage.test.ts
@@ -7,7 +7,7 @@ import { v4 as uuidv4 } from 'uuid'
 import Logger from '../../utils/Logger'
 import crypto from 'crypto'
 import elliptic from 'elliptic'
-import { ChallengeReactNativeStorage } from '../ChallengeStorage'
+import { ReactNativeChallengeStorage } from '../react-native/ReactNativeChallengeStorage'
 const ecl = new elliptic.ec('p256')
 
 const completeFlowWithServer = async () => {
@@ -15,7 +15,7 @@ const completeFlowWithServer = async () => {
     email: `test-${uuidv4()}@test.usecapsule.com`,
   })
   await userManagementClient.verifyEmail(userId, { verificationCode: '123456' })
-  const storage = new ChallengeReactNativeStorage(userId)
+  const storage = new ReactNativeChallengeStorage(userId)
 
   await userManagementClient.addBiometrics(userId, { publicKey: await storage.getPublicKey() })
   const challenge = await userManagementClient.getBiometricsChallenge(userId)
@@ -33,7 +33,7 @@ const completeFlowWithServer = async () => {
 }
 
 const completeFlowOffline = async () => {
-  const storage = new ChallengeReactNativeStorage('123')
+  const storage = new ReactNativeChallengeStorage('123')
   const message = '1d52c368-d91c-46f4-b449-fa142c8b812d'
   void (await storage.getPublicKey()) // we cannot sign without biometric initialized.
   const signature = await storage.signChallenge(message)
