+++ src/web3/contracts.ts
@@ -11,7 +11,7 @@ import { call, select } from 'redux-saga/effects'
 import { ContractKitEvents } from 'src/analytics/Events'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { ErrorMessages } from 'src/app/ErrorMessages'
-import { CapsuleWallet } from 'src/capsule/CapsuleWallet'
+import { CapsuleWallet } from 'src/capsule/react-native/ReactNativeCapsuleWallet'
 import { DEFAULT_FORNO_URL } from 'src/config'
 //import { isProviderConnectionError } from 'src/geth/geth'
 //import { GethNativeBridgeWallet } from 'src/geth/GethNativeBridgeWallet'
@@ -22,7 +22,7 @@ import { getHttpProvider, getIpcProvider } from 'src/web3/providers'
 import { fornoSelector } from 'src/web3/selectors'
 import Web3 from 'web3'
 import { EIP712TypedData } from '@celo/utils/lib/sign-typed-data-utils'
-import { ReactNativeSignersStorage } from '../capsule/SignersStorage'
+import { ReactNativeSignersStorage } from '../capsule/react-native/ReactNativeSignersStorage'
 
 const TAG = 'web3/contracts'
 // const KIT_INIT_RETRY_DELAY = 2000
